UPDATE moduleList Button
apply useState to modulelist button
