+++ frontend/src/components/notification/moduleList/moduleTest.ts
@@ -1,0 +1,7 @@
+export const MODULES = [
+    { id: 0, name: "All", isSelect: false },
+    { id: 1, name: "Shop", isSelect: false },
+    { id: 2, name: "TodoList", isSelect: false },
+    { id: 3, name: "Chat", isSelect: true },
+    { id: 3, name: "Transaction", isSelect: false },
+]
